RTL support! (to some extent...) -Shoutout: MiladJ
All the credits go for MiladJ for implementing this in the Farsi/Persian Repo.
https://github.com/progit2-fa/progit2/pull/14/commits/e3f695b6bfed1c9ce2eb7f2df2b46071ce11f2a0

Updated Gemfile to the exact versions that Milad used.
Added a DocInfo and activated it.

Still has some flaws tho and I keep working on it to perfect it. But for
now I couldn't keep this any longer to myself and my repo.